refactor: login, register view models, and typos - moved login, register and user viewmodels to viewmodels package - fixed hardcoded username display - fixed minor typos - added automatic username creation in register screen, concatenating first name and surname
