+++ app/src/main/java/com/pgmv/bandify/ui/screen/LoginScreen.kt
@@ -1,7 +1,6 @@
 package com.pgmv.bandify.ui.screen
 
 
-import android.util.Patterns
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
@@ -18,83 +17,11 @@ import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import androidx.navigation.NavHostController
 import com.pgmv.bandify.R
-import com.pgmv.bandify.database.DatabaseHelper
-import kotlinx.coroutines.launch
+import com.pgmv.bandify.viewmodel.AuthenticationViewModel
+import com.pgmv.bandify.viewmodel.UserViewModel
 
-
-class AuthenticationViewModel(private val dbHelper: DatabaseHelper) : ViewModel() {
-
-    var email by mutableStateOf("")
-    var password by mutableStateOf("")
-    var emailError by mutableStateOf<String?>(null)
-    var passwordError by mutableStateOf<String?>(null)
-    var loginError by mutableStateOf<String?>(null)
-    var loggedInUserId by mutableStateOf<Long?>(null)
-    var wasLoginAttempted by mutableStateOf(false)
-
-    fun updateEmail(input: String) {
-        email = input
-        if (wasLoginAttempted) {
-            validateFields()
-        }
-    }
-
-    fun updatePassword(input: String) {
-        password = input
-        if (wasLoginAttempted) {
-            validateFields()
-        }
-    }
-
-    fun validateFields(): Boolean {
-        emailError = when {
-            email.isEmpty() -> "Campo obrigatório"
-            !Patterns.EMAIL_ADDRESS.matcher(email).matches() -> "Formato de e-mail inválido"
-            else -> null
-        }
-
-        passwordError = when {
-            password.isEmpty() -> "Campo obrigatório"
-            password.length < 6 -> "A senha deve ter pelo menos 6 caracteres"
-            else -> null
-        }
-
-        return emailError == null && passwordError == null
-    }
-
-    fun login(onSuccess: (Long) -> Unit, onError: (String) -> Unit) {
-        wasLoginAttempted = true
-        if (!validateFields()) {
-            loginError = "Por favor, corrija os campos destacados."
-            return
-        }
-
-        viewModelScope.launch {
-            try {
-                val user = dbHelper.userDao().validateUser(email, password)
-                if (user != null) {
-                    loggedInUserId = user.id
-                    onSuccess(user.id)
-                } else {
-                    loginError = "Usuário ou senha inválidos."
-                    onError("Usuário ou senha inválidos.")
-                }
-            } catch (e: Exception) {
-                loginError = "Erro ao conectar ao banco de dados."
-                onError("Erro ao conectar ao banco de dados.")
-            }
-        }
-    }
-}
-
-
-class UserViewModel : ViewModel() {
-    var userId by mutableStateOf<Long?>(null)
-}
 
 @Composable
 fun LoginScreen(
