+++ app/src/main/java/com/pgmv/bandify/ui/screen/RegisterScreen.kt
@@ -41,6 +41,7 @@ import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.text.toLowerCase
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
@@ -50,59 +51,13 @@ import androidx.navigation.NavHostController
 import com.pgmv.bandify.R
 import com.pgmv.bandify.database.DatabaseHelper
 import com.pgmv.bandify.domain.User
+import com.pgmv.bandify.viewmodel.RegisterViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 
-class RegisterViewModel : ViewModel() {
 
-    var firstName by mutableStateOf("")
-        private set
-
-    var surname by mutableStateOf("")
-        private set
-
-    var email by mutableStateOf("")
-        private set
-
-    var password by mutableStateOf("")
-        private set
-
-    var confirmPassword by mutableStateOf("")
-        private set
-
-    var firstNameHasErrors by mutableStateOf(false)
-    var surnameHasErrors by mutableStateOf(false)
-    var emailHasErrors by mutableStateOf(false)
-    var passwordHasErrors by mutableStateOf(false)
-    var confirmPasswordHasErrors by mutableStateOf(false)
-
-    fun updatefirstName(input: String) {
-        firstName = input
-        firstNameHasErrors = firstName.isEmpty()
-    }
-
-    fun updateSurname(input: String) {
-        surname = input
-        surnameHasErrors = surname.isEmpty()
-    }
-
-    fun updateEmail(input: String) {
-        email = input
-        emailHasErrors = email.isEmpty() || !Patterns.EMAIL_ADDRESS.matcher(email).matches()
-    }
-
-    fun updatePassword(input: String) {
-        password = input
-        passwordHasErrors = password.isEmpty() || password.length < 6
-    }
-
-    fun updateConfirmPassword(input: String) {
-        confirmPassword = input
-        confirmPasswordHasErrors = confirmPassword.isEmpty() || confirmPassword != password
-    }
-}
 
 @Composable
 fun RegisterScreen(navController: NavHostController) {
@@ -175,7 +130,7 @@ fun RegisterScreen(navController: NavHostController) {
             }
 
             Text(
-                text = "SEU SEGUNDO NOME",
+                text = "SEU SOBRENOME",
                 fontSize = 12.sp,
                 color = MaterialTheme.colorScheme.primary
             )
@@ -333,8 +288,8 @@ fun RegisterScreen(navController: NavHostController) {
                                 email = viewModel.email,
                                 password = viewModel.password,
                                 surname = viewModel.surname,
-                                username = "Efraim",
-                                phone = "784781818",
+                                username = (viewModel.firstName + viewModel.surname).lowercase(),
+                                phone = "",
                             )
                             val rowId = dbHelper.userDao().insertUser(newUser)
                             withContext(Dispatchers.Main) {
