+++ app/src/main/java/com/pgmv/bandify/viewmodel/UserViewModel.kt
@@ -1,0 +1,10 @@
+package com.pgmv.bandify.viewmodel
+
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
+import androidx.lifecycle.ViewModel
+
+class UserViewModel : ViewModel() {
+    var userId by mutableStateOf<Long?>(null)
+}
